+++ src/Entity/Basket.php
@@ -97,7 +97,7 @@ class Basket
 
     public function __toString()
     {
-        return $this->number;
+        return $this->number ?? '';
     }
 
     public function getId(): ?int