+++ src/Service/ProductServiceInterface.php
@@ -12,6 +12,8 @@ interface ProductServiceInterface
 
     public function findAllPaginated($query);
 
+    public function findAllMedias();
+
     public function findOneById(int $id): Product;
 
     public function search(string $query);