+++ src/Service/ShopService.php
@@ -2,15 +2,17 @@
 
 namespace c975L\ShopBundle\Service;
 
-use c975L\ShopBundle\Entity\Shop;
-use c975L\ShopBundle\Repository\ProductRepository;
+use DateTimeImmutable;
+use Doctrine\ORM\EntityManagerInterface;
 use Knp\Component\Pager\PaginatorInterface;
-
+use c975L\ShopBundle\Entity\ProductItemDownload;
+use c975L\ShopBundle\Repository\ProductRepository;
 
 class ShopService implements ShopServiceInterface
 {
     public function __construct(
         private readonly ProductRepository $productRepository,
+        private readonly EntityManagerInterface $entityManager,
         private readonly PaginatorInterface $paginator
     ) {
     }