+++ src/Service/ShopServiceInterface.php
@@ -3,6 +3,7 @@
 namespace c975L\ShopBundle\Service;
 
 use c975L\ShopBundle\Entity\Shop;
+use c975L\ShopBundle\Entity\ProductItemDownload;
 
 interface ShopServiceInterface
 {